create exercise thumbnail
